Support boards with a voltage divider on VBUS
This can be useful to allow measuring the actual USB voltage, rather
than just detect presence.

Previously the USB code would always enable an internal pulldown on the
VBUS, which would skew the reading (and, if the divider is fairly high
impedance, can even prevent detecting VBUS presence using a digital
read).

With this commit, boards can indicate they use an external divider using
the STM32L0_CONFIG_PIN_VBUS_HAS_DIVIDER macro in their variant.h. When
this is defined, the internal pullup is not enabled to prevent it from
causing issues and because it is no longer needed (the external divider
functions as a pulldown).